+++ spec/lumberjack/logger_spec.rb
@@ -52,17 +52,28 @@ RSpec.describe Lumberjack::Logger do
 
     it "should open a file path in a device" do
       logger = Lumberjack::Logger.new(File.join(tmp_dir, "log_file_1.log"))
-      expect(logger.device.class).to eq(Lumberjack::Device::LogFile)
+      expect(logger.device.class).to eq(Lumberjack::Device::File)
     end
 
     it "should open a pathname in a device" do
       logger = Lumberjack::Logger.new(Pathname.new(File.join(tmp_dir, "log_file_1.log")))
-      expect(logger.device.class).to eq(Lumberjack::Device::LogFile)
+      expect(logger.device).to be_a(Lumberjack::Device::File)
+    end
+
+    it "should open a File in a device" do
+      file = File.new(File.join(tmp_dir, "log_file_1.log"), "w")
+      logger = Lumberjack::Logger.new(file)
+      expect(logger.device.class).to eq(Lumberjack::Device::File)
     end
 
     it "should use the null device if the stream is :null" do
       logger = Lumberjack::Logger.new(:null)
       expect(logger.device.class).to eq(Lumberjack::Device::Null)
+    end
+
+    it "should use the test device if the stream is :test" do
+      logger = Lumberjack::Logger.new(:test)
+      expect(logger.device.class).to eq(Lumberjack::Device::Test)
     end
 
     it "should set the level with a numeric" do